requirements.txt
streamlit
dotenv
fpdf
google-generativeai
